experiments/004-world-model-induction-v4/pyproject.toml
[project]
name = "experiment-004-v3"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "arc-agi>=0.9.4",
    "openai>=1.72.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.2",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
]
